+++ kong/plugins/prometheus/exporter.lua
@@ -152,21 +152,21 @@ local function init()
   if http_subsystem then
     metrics.status = prometheus:counter("http_requests_total",
                                         "HTTP status codes per consumer/service/route in Kong",
-                                        {"service", "route", "code", "source", "consumer"})
+                                        {"service", "route", "code", "source", "tags", "consumer"})
   else
     metrics.status = prometheus:counter("stream_sessions_total",
                                         "Stream status codes per service/route in Kong",
-                                        {"service", "route", "code", "source"})
+                                        {"service", "route", "code", "source", "tags"})
   end
   metrics.kong_latency = prometheus:histogram("kong_latency_ms",
                                               "Latency added by Kong and enabled plugins " ..
                                               "for each service/route in Kong",
-                                              {"service", "route"},
+                                              {"service", "route", "tags"},
                                               KONG_LATENCY_BUCKETS)
   metrics.upstream_latency = prometheus:histogram("upstream_latency_ms",
                                                   "Latency added by upstream response " ..
                                                   "for each service/route in Kong",
-                                                  {"service", "route"},
+                                                  {"service", "route", "tags"},
                                                   UPSTREAM_LATENCY_BUCKETS)
 
 
@@ -174,13 +174,13 @@ local function init()
     metrics.total_latency = prometheus:histogram("request_latency_ms",
                                                  "Total latency incurred during requests " ..
                                                  "for each service/route in Kong",
-                                                 {"service", "route"},
+                                                 {"service", "route", "tags"},
                                                  UPSTREAM_LATENCY_BUCKETS)
   else
     metrics.total_latency = prometheus:histogram("session_duration_ms",
                                                  "latency incurred in stream session " ..
                                                  "for each service/route in Kong",
-                                                 {"service", "route"},
+                                                 {"service", "route", "tags"},
                                                  UPSTREAM_LATENCY_BUCKETS)
   end
 
@@ -188,12 +188,12 @@ local function init()
     metrics.bandwidth = prometheus:counter("bandwidth_bytes",
                                           "Total bandwidth (ingress/egress) " ..
                                           "throughput in bytes",
-                                          {"service", "route", "direction", "consumer"})
+                                          {"service", "route", "direction", "tags", "consumer"})
   else -- stream has no consumer
     metrics.bandwidth = prometheus:counter("bandwidth_bytes",
                                           "Total bandwidth (ingress/egress) " ..
                                           "throughput in bytes",
-                                          {"service", "route", "direction"})
+                                          {"service", "route", "direction", "tags"})
   end
 
   -- Hybrid mode status
@@ -243,9 +243,9 @@ end
 
 -- Since in the prometheus library we create a new table for each diverged label
 -- so putting the "more dynamic" label at the end will save us some memory
-local labels_table_bandwidth = {0, 0, 0, 0}
-local labels_table_status = {0, 0, 0, 0, 0}
-local labels_table_latency = {0, 0}
+local labels_table_bandwidth = {0, 0, 0, 0, 0}
+local labels_table_status = {0, 0, 0, 0, 0, 0}
+local labels_table_latency = {0, 0, 0}
 local upstream_target_addr_health_table = {
   { value = 0, labels = { 0, 0, 0, "healthchecks_off", ngx.config.subsystem } },
   { value = 0, labels = { 0, 0, 0, "healthy", ngx.config.subsystem } },
@@ -297,7 +297,8 @@ local function log(message, serialized)
   if serialized.ingress_size or serialized.egress_size then
     labels_table_bandwidth[1] = service_name
     labels_table_bandwidth[2] = route_name
-    labels_table_bandwidth[4] = consumer
+    labels_table_bandwidth[4] = serialized.tags
+    labels_table_bandwidth[5] = consumer
 
     local ingress_size = serialized.ingress_size
     if ingress_size and ingress_size > 0 then
@@ -316,6 +317,7 @@ local function log(message, serialized)
     labels_table_status[1] = service_name
     labels_table_status[2] = route_name
     labels_table_status[3] = serialized.status_code
+    labels_table_status[5] = serialized.tags
 
     if kong.response.get_source() == "service" then
       labels_table_status[4] = "service"
@@ -323,7 +325,7 @@ local function log(message, serialized)
       labels_table_status[4] = "kong"
     end
 
-    labels_table_status[5] = consumer
+    labels_table_status[6] = consumer
 
     metrics.status:inc(1, labels_table_status)
   end
@@ -331,6 +333,7 @@ local function log(message, serialized)
   if serialized.latencies then
     labels_table_latency[1] = service_name
     labels_table_latency[2] = route_name
+    labels_table_latency[3] = serialized.tags
 
     if http_subsystem then
       local request_latency = serialized.latencies.request