+++ kong/plugins/prometheus/handler.lua
@@ -2,6 +2,9 @@ local exporter = require "kong.plugins.prometheus.exporter"
 local kong = kong
 local kong_meta = require "kong.meta"
 
+local keys = require('pl.tablex').keys
+local sort = table.sort
+local concat = table.concat
 
 exporter.init()
 
@@ -33,6 +36,26 @@ function PrometheusHandler.log(self, conf)
       serialized.status_code = message.session.status
     end
   end
+
+  local tags = {}
+  if conf.expose_tags.from_service and message.service and message.service.tags then
+    for i = 1, #message.service.tags do
+      tags[message.service.tags[i]] = true
+    end
+  end
+  if conf.expose_tags.from_route and message.route and message.route.tags then
+    for i = 1, #message.route.tags do
+      tags[message.route.tags[i]] = true
+    end
+  end
+  if conf.expose_tags.from_consumer and message.consumer and message.consumer.tags then
+    for i = 1, #message.consumer.tags do
+      tags[message.consumer.tags[i]] = true
+    end
+  end
+  tags = keys(tags)
+  sort(tags)
+  serialized.tags = concat(tags, ",")
 
   if conf.bandwidth_metrics then
     if http_subsystem then